+++ examples/demo/public/index.tsx
@@ -1,6 +1,4 @@
-import { LocationProvider, Router } from 'preact-iso/router';
-import lazy, { ErrorBoundary } from 'preact-iso/lazy';
-import hydrate from 'preact-iso/hydrate';
+import { LocationProvider, Router, lazy, ErrorBoundary, hydrate } from 'preact-iso';
 import Home from './pages/home.js';
 // import About from './pages/about/index.js';
 import NotFound from './pages/_404.js';
@@ -9,13 +7,17 @@ import Header from './header.tsx';
 
 const sleep = t => new Promise(r => setTimeout(r, t));
 
+const IS_CLIENT = typeof window !== 'undefined';
+
 const About = lazy(() => import('./pages/about/index.js'));
-const LazyAndLate = lazy(async () => (await sleep(1500), import('./pages/about/index.js')));
+const LazyAndLate = lazy(async () => (IS_CLIENT && (await sleep(1500)), import('./pages/about/index.js')));
 const CompatPage = lazy(() => import('./pages/compat.js'));
-const ClassFields = lazy(async () => (await sleep(1500), import('./pages/class-fields.js')));
+const ClassFields = lazy(async () => (IS_CLIENT && (await sleep(1500)), import('./pages/class-fields.js')));
 const Files = lazy(() => import('./pages/files/index.js'));
 const Environment = lazy(async () => (await import('./pages/environment/index.js')).Environment);
 const JSONView = lazy(async () => (await import('./pages/json.js')).JSONView);
+const MetaTags = lazy(async () => (await import('./pages/meta-tags.js')).MetaTags);
+const AliasOutside = lazy(async () => await import('./pages/alias-outside.js'));
 
 function showLoading() {
 	document.body.classList.add('loading');
@@ -39,6 +41,8 @@ export function App() {
 						<Files path="/files" />
 						<Environment path="/env" />
 						<JSONView path="/json" />
+						<MetaTags path="/meta-tags" />
+						<AliasOutside path="/alias-outside" />
 						<NotFound default />
 					</Router>
 				</ErrorBoundary>
@@ -52,8 +56,7 @@ if (typeof window !== 'undefined') {
 }
 
 export async function prerender(data) {
-	const { default: render } = await import('preact-iso/prerender');
-	return await render(<App {...data} />);
+	return (await import('./prerender.js')).prerender(<App {...data} />);
 }
 
 // @ts-ignore