+++ packages/wmr/src/lib/prerender.js
@@ -50,33 +50,12 @@ async function workerCode({ cwd, out, publicPath }) {
 
 	const path = require('path');
 	const fs = require('fs').promises;
-	const BEFORE = Symbol('before');
-	const AFTER = Symbol('after');
 
 	function enc(str) {
 		return str.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
 	}
 
 	globalThis.location = /** @type {object} */ ({});
-
-	globalThis.document = /** @type {object} */ ({
-		createElement(type) {
-			return { type };
-		},
-		querySelector() {},
-		head: {
-			[BEFORE]: '',
-			[AFTER]: '',
-			children: /** @type {any[]} */ ([]),
-			appendChild(c) {
-				this.children.push(c);
-			},
-			/** @param {'afterbegin'|'beforeend'} position @param {string} html */
-			insertAdjacentHTML(position, html) {
-				this[position === 'afterbegin' ? BEFORE : AFTER] += String(html);
-			}
-		}
-	});
 
 	globalThis.self = /** @type {any} */ (globalThis);
 
@@ -90,14 +69,31 @@ async function workerCode({ cwd, out, publicPath }) {
 	// Grab the generated HTML file, which we'll use as a template:
 	const tpl = await fs.readFile(path.resolve(cwd, out, 'index.html'), 'utf-8');
 
-	// The first script in the file is assumed to have a .prerender export:
-	let script = (tpl.match(/<script(?:\s[^>]*?)?\s+src=(['"]?)([^>]*?)\1(?:\s[^>]*?)?>/) || [])[2];
+	// The first script in the file that is not external is assumed to have a
+	// `prerender` export
+	let script;
+	const SCRIPT_TAG = /<script(?:\s[^>]*?)?\s+src=(['"]?)([^>]*?)\1(?:\s[^>]*?)?>/g;
+
+	let match;
+	while ((match = SCRIPT_TAG.exec(tpl))) {
+		// Ignore external urls
+		if (!match || /^(?:https?|file|data)/.test(match[2])) continue;
+
+		script = match[2].replace(publicPath, '').replace(/^(\.?\/)?/g, '');
+		script = path.resolve(cwd, out, script);
+	}
+
 	if (!script) {
 		throw Error(`Unable to detect <script src="entry.js"> in your index.html.`);
 	}
-	// script = new URL(`../dist/${script.replace(/^(\.?\/)?/g, '')}`, selfUrl).href;
-	script = script.replace(publicPath, '');
-	script = path.resolve(cwd, out, script.replace(/^(\.?\/)?/g, ''));
+
+	/** @typedef {{ type: string, props: Record<string, string>, children?: string } | string | null} HeadElement */
+
+	/**
+	 * @type {{ lang: string, title: string, elements: Set<HeadElement>}}
+	 */
+	let head = { lang: '', title: '', elements: new Set() };
+	globalThis.wmr = { ssr: { head } };
 
 	// Prevent Rollup from transforming `import()` here.
 	const $import = new Function('s', 'return import(s)');
@@ -105,6 +101,32 @@ async function workerCode({ cwd, out, publicPath }) {
 	const doPrerender = m.prerender;
 	// const App = m.default || m[Object.keys(m)[0]];
 
+	/**
+	 * @param {HeadElement|HeadElement[]|Set<HeadElement>} element
+	 * @returns {string} html
+	 */
+	function serializeElement(element) {
+		if (element == null) return '';
+		if (typeof element !== 'object') return String(element);
+		if (Array.isArray(element)) return element.map(serializeElement).join('');
+		const type = element.type;
+		let s = `<${type}`;
+		const props = element.props || {};
+		let children = element.children;
+		for (const prop of Object.keys(props).sort()) {
+			const value = props[prop];
+			// Filter out empty values:
+			if (value == null) continue;
+			if (prop === 'children' || prop === 'textContent') children = value;
+			else s += ` ${prop}="${enc(value)}"`;
+		}
+		s += '>';
+		if (!/link|meta|base/.test(type)) {
+			if (children) s += serializeElement(children);
+			s += `</${type}>`;
+		}
+		return s;
+	}
 	// We start by pre-rendering the homepage.
 	// Links discovered during pre-rendering get pushed into the list of routes.
 	const seen = new Set(['/']);
@@ -125,10 +147,7 @@ async function workerCode({ cwd, out, publicPath }) {
 			} catch {}
 		}
 
-		// Reset document.head so that CSS for the current route will be injected into it:
-		// @ts-ignore
-		const head = (document.head.children = []);
-		document.head[BEFORE] = document.head[AFTER] = '';
+		head = { lang: '', title: '', elements: new Set() };
 
 		// Do pre-rendering, as defined by the entry chunk:
 		const result = await doPrerender({ ssr: true, url: route.url, route });
@@ -146,25 +165,45 @@ async function workerCode({ cwd, out, publicPath }) {
 				routes.push({ url, _discoveredBy: route });
 			}
 		}
-		const body = (result && result.html) || result;
+
+		let body;
+		if (result && typeof result === 'object') {
+			if (result.html) body = result.html;
+			if (result.head) {
+				head = result.head;
+			}
+
+			if (result.data && typeof result.data === 'object') {
+				body += `<script type="isodata">${JSON.stringify(result.data)}</script>`;
+			} else if (result.data) {
+				console.warn('You passed in prerender-data in a non-object format: ', result.data);
+			}
+		} else {
+			body = result;
+		}
+
+		// TODO: Use a proper HTML parser here. We should definitely not parse HTML
+		// with regex :S
 
 		// Inject HTML links at the end of <head> for any stylesheets injected during rendering of the page:
-		let headHtml = [
-			document.head[BEFORE],
-			...new Set(head.filter(c => c.rel && c.href).map(c => `<link rel="${enc(c.rel)}" href="${enc(c.href)}">`)),
-			document.head[AFTER]
-		].join('');
+		let headHtml = head.elements ? Array.from(new Set(Array.from(head.elements).map(serializeElement))).join('') : '';
 
 		let html = tpl;
 
-		if (document.title) {
-			const title = `<title>${enc(document.title)}</title>`;
+		if (head.title) {
+			const title = `<title>${enc(head.title)}</title>`;
 			const matchTitle = /<title>([^<>]*?)<\/title>/i;
 			if (matchTitle.test(html)) {
 				html = html.replace(matchTitle, title);
 			} else {
 				headHtml = title + headHtml;
 			}
+		}
+
+		if (head.lang) {
+			// TODO: This removes any existing attributes, but merging them without
+			// a proper HTML parser is way too error prone.
+			html = html.replace(/(<html(\s[^>]*?)?>)/, `<html lang="${enc(head.lang)}">`);
 		}
 
 		html = html.replace(/(<\/head>)/, headHtml + '$1');