+++ packages/wmr/src/plugins/wmr/styles-plugin.js
@@ -1,6 +1,8 @@
 import { promises as fs } from 'fs';
 import { basename, dirname, relative, resolve, sep, posix } from 'path';
+import { transformCssImports } from '../../lib/transform-css-imports.js';
 import { transformCss } from '../../lib/transform-css.js';
+import { matchAlias } from '../../lib/aliasing.js';
 
 /**
  * @param {string} sass
@@ -91,14 +93,19 @@ export async function modularizeCss(css, id, mappings = [], idAbsolute) {
 
 /**
  * Implements hot-reloading for stylesheets imported by JS.
- * @param {object} [options]
+ * @param {object} options
  * @param {string} [options.cwd] Manually specify the cwd from which to resolve filenames (important for calculating hashes!)
  * @param {boolean} [options.hot] Indicates the plugin should inject a HMR-runtime
  * @param {boolean} [options.fullPath] Preserve the full original path when producing CSS assets
+ * @param {boolean} [options.production]
+ * @param {Record<string, string>} options.aliases
  * @returns {import('rollup').Plugin}
  */
-export default function wmrStylesPlugin({ cwd, hot, fullPath } = {}) {
+export default function wmrStylesPlugin({ cwd, hot, fullPath, production, aliases }) {
 	const cwds = new Set();
+
+	let assetId = 0;
+	const assetMap = new Map();
 
 	return {
 		name: 'wmr-styles',
@@ -117,7 +124,14 @@ export default function wmrStylesPlugin({ cwd, hot, fullPath } = {}) {
 			const isIcss = /(composes:|:global|:local)/.test(source);
 			const isModular = /\.module\.(css|s[ac]ss)$/.test(id);
 
-			let idRelative = cwd ? relative(cwd || '', resolve(cwd, id)) : multiRelative(cwds, id);
+			let idRelative = id;
+			let aliased = matchAlias(aliases, id);
+			if (aliased) {
+				idRelative = aliased.slice('/@alias/'.length);
+			} else {
+				idRelative = cwd ? relative(cwd || '', id) : multiRelative(cwds, id);
+			}
+
 			if (idRelative.match(/^[^/]*\\/)) idRelative = idRelative.split(sep).join(posix.sep);
 
 			const mappings = [];
@@ -130,10 +144,37 @@ export default function wmrStylesPlugin({ cwd, hot, fullPath } = {}) {
 				source = transformCss(source);
 			}
 
+			// Note: `plugin.generateBundle` is only called during prod builds for
+			// CSS files. So we need to guard the url replacement code.
+			if (production) {
+				source = await transformCssImports(source, idRelative, {
+					async resolveId(spec) {
+						// Rollup doesn't allow assets to depend on other assets. This is a
+						// pretty common case for CSS files referencing images or fonts. To
+						// work around that we'll rewrite every file reference to
+						// `__WMR_ASSET_ID_##`, similar to `import.meta.ROLLUP_FILE_URL_##`.
+						// We'll resolve those in the `generateBundle` phase.
+						// See: https://github.com/rollup/rollup/issues/2872
+						if (spec.indexOf(':') === -1) {
+							const absolute = resolve(dirname(idRelative), spec.split(posix.sep).join(sep));
+
+							if (!absolute.startsWith(cwd)) return;
+
+							const ref = `__WMR_ASSET_ID_${assetId++}`;
+							assetMap.set(ref, {
+								filePath: absolute,
+								source: absolute.endsWith('.css') ? await fs.readFile(absolute, 'utf-8') : await fs.readFile(absolute)
+							});
+							return ref;
+						}
+					}
+				});
+			}
+
 			const ref = this.emitFile({
 				type: 'asset',
 				name: fullPath ? undefined : basename(id).replace(/\.s[ac]ss$/, '.css'),
-				fileName: fullPath ? idRelative : undefined,
+				fileName: fullPath ? (aliased ? `@alias/${idRelative}` : idRelative) : undefined,
 				source
 			});
 
@@ -177,6 +218,24 @@ export default function wmrStylesPlugin({ cwd, hot, fullPath } = {}) {
 				syntheticNamedExports: true,
 				map: null
 			};
+		},
+		async generateBundle(_, bundle) {
+			for (const id in bundle) {
+				const item = bundle[id];
+
+				if (item.type === 'asset' && item.fileName.endsWith('.css')) {
+					item.source = item.source.replace(/__WMR_ASSET_ID_\d+/g, m => {
+						const mapped = assetMap.get(m);
+						const ref = this.emitFile({
+							type: 'asset',
+							name: basename(mapped.filePath),
+							source: mapped.source
+						});
+
+						return '/' + this.getFileName(ref);
+					});
+				}
+			}
 		}
 	};
 }