+++ packages/wmr/test/production.test.js
@@ -1,6 +1,6 @@
 import path from 'path';
 import { promises as fs } from 'fs';
-import { setupTest, teardown, runWmr, loadFixture, serveStatic } from './test-helpers.js';
+import { setupTest, teardown, runWmr, loadFixture, serveStatic, withLog } from './test-helpers.js';
 import { printCoverage, analyzeTrace } from './tracing-helpers.js';
 
 jest.setTimeout(30000);
@@ -43,6 +43,18 @@ describe('production', () => {
 		expect(await env.page.content()).toMatch(/foobarbaz/);
 	});
 
+	it('should throw error on missing module type', async () => {
+		await loadFixture('script-type', env);
+		instance = await runWmr(env.tmp.path, 'build');
+		await withLog(instance.output, async () => {
+			const code = await instance.done;
+			expect(code).toEqual(1);
+
+			const log = instance.output.join('\n');
+			expect(log).toMatch(/No module scripts were found/);
+		});
+	});
+
 	it('should allow overwriting url loader', async () => {
 		await loadFixture('overwrite-loader-url', env);
 		instance = await runWmr(env.tmp.path, 'build');
@@ -63,6 +75,134 @@ describe('production', () => {
 		expect(text).toMatch(/my-url: \/assets\/foo\..*\.svg/);
 		expect(text).toMatch(/url: \/assets\/foo\..*\.svg/);
 		expect(text).toMatch(/fallback: \/assets\/foo\..*\.svg/);
+	});
+
+	it('should show all generated files in cli output', async () => {
+		await loadFixture('file-import', env);
+		instance = await runWmr(env.tmp.path, 'build');
+		const code = await instance.done;
+		const output = instance.output;
+		console.log(output);
+
+		expect(code).toEqual(0);
+
+		const stats = output.slice(output.findIndex(line => /Wrote.*to disk/.test(line)));
+		expect(stats.join('\n')).toMatch(/img\..*\.jpg/);
+	});
+
+	it('should support virtual ids', async () => {
+		await loadFixture('virtual-id', env);
+		instance = await runWmr(env.tmp.path, 'build');
+		const code = await instance.done;
+		expect(code).toEqual(0);
+
+		const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+		cleanup.push(stop);
+
+		await env.page.goto(address, {
+			waitUntil: ['networkidle0', 'load']
+		});
+
+		const text = await env.page.content();
+		expect(text).toMatch(/it works/);
+	});
+
+	describe('alias', () => {
+		it('should alias directories', async () => {
+			await loadFixture('alias-outside', env);
+			instance = await runWmr(env.tmp.path, 'build');
+			const code = await instance.done;
+			await withLog(instance.output, async () => {
+				expect(code).toEqual(0);
+
+				const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+				cleanup.push(stop);
+
+				await env.page.goto(address, {
+					waitUntil: ['networkidle0', 'load']
+				});
+
+				const text = await env.page.content();
+				expect(text).toMatch(/it works/);
+			});
+		});
+
+		it('should alias src by default', async () => {
+			await loadFixture('alias-src', env);
+			instance = await runWmr(env.tmp.path, 'build');
+			const code = await instance.done;
+			await withLog(instance.output, async () => {
+				expect(code).toEqual(0);
+
+				const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+				cleanup.push(stop);
+
+				await env.page.goto(address, {
+					waitUntil: ['networkidle0', 'load']
+				});
+
+				const text = await env.page.content();
+				expect(text).toMatch(/it works/);
+			});
+		});
+
+		it('should alias CSS', async () => {
+			await loadFixture('alias-css', env);
+			instance = await runWmr(env.tmp.path, 'build');
+			const code = await instance.done;
+			await withLog(instance.output, async () => {
+				expect(code).toEqual(0);
+
+				const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+				cleanup.push(stop);
+
+				await env.page.goto(address, {
+					waitUntil: ['networkidle0', 'load']
+				});
+
+				const color = await env.page.$eval('h1', el => getComputedStyle(el).color);
+				expect(color).toBe('rgb(255, 218, 185)');
+			});
+		});
+	});
+
+	describe('CSS', () => {
+		it('should resolve CSS imports', async () => {
+			await loadFixture('css-imports', env);
+			instance = await runWmr(env.tmp.path, 'build');
+			const code = await instance.done;
+			expect(code).toEqual(0);
+
+			const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+			cleanup.push(stop);
+
+			await env.page.goto(address, {
+				waitUntil: ['networkidle0', 'load']
+			});
+
+			expect(await env.page.$eval('h1', el => getComputedStyle(el).color)).toBe('rgb(255, 0, 0)');
+			expect(await env.page.$eval('h1', el => getComputedStyle(el).backgroundColor)).toBe('rgb(255, 218, 185)');
+			expect(await env.page.$eval('body', el => getComputedStyle(el).backgroundImage)).toContain('.jpg');
+		});
+	});
+
+	describe('import.meta.env', () => {
+		it('should support process.env.NODE_ENV', async () => {
+			await loadFixture('import-meta-env', env);
+			instance = await runWmr(env.tmp.path, 'build');
+			const code = await instance.done;
+
+			expect(code).toEqual(0);
+
+			const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+			cleanup.push(stop);
+
+			await env.page.goto(address, {
+				waitUntil: ['networkidle0', 'load']
+			});
+
+			expect(await env.page.content()).toMatch(/production/);
+		});
 	});
 
 	describe('demo app', () => {
@@ -257,6 +397,28 @@ describe('production', () => {
 			const index = await readfile('dist/' + dist.find(f => f.match(/^index\.\w+\.js$/)));
 			expect(index).toContain(`("/assets/${assets[0]}")`);
 		});
+
+		it('should hoist entry CSS into HTML <link> tag', async () => {
+			await loadFixture('css-entry', env);
+			instance = await runWmr(env.tmp.path, 'build');
+			const code = await instance.done;
+
+			await withLog(instance.output, async () => {
+				expect(code).toBe(0);
+
+				const files = (await fs.readdir(path.join(env.tmp.path, 'dist', 'assets'))).filter(f => f[0] !== '.');
+				const css = files.find(f => f.endsWith('.css'));
+
+				const { address, stop } = serveStatic(path.join(env.tmp.path, 'dist'));
+				cleanup.push(stop);
+
+				await env.page.goto(address, {
+					waitUntil: ['networkidle0', 'load']
+				});
+
+				expect(await env.page.content()).toContain(`<link rel="stylesheet" href="/assets/${css}">`);
+			});
+		});
 	});
 
 	describe('config.publicPath', () => {
@@ -343,6 +505,37 @@ describe('production', () => {
 		 */
 		const readdir = async (env, f) => (await fs.readdir(path.join(env.tmp.path, f))).filter(f => f[0] !== '.');
 
+		it('should print warning on missing entry', async () => {
+			await loadFixture('prerender-missing-entry', env);
+			instance = await runWmr(env.tmp.path, 'build', '--prerender');
+			const code = await instance.done;
+			expect(instance.output.join('\n')).toMatch(/file not found/i);
+			expect(instance.output.join('\n')).toMatch(/is the extension correct/i);
+			expect(code).toBe(1);
+		});
+
+		it('should not try to prerender external scripts', async () => {
+			await loadFixture('prerender-external', env);
+			instance = await runWmr(env.tmp.path, 'build', '--prerender');
+			const code = await instance.done;
+			expect(instance.output.join('\n')).toMatch(/Prerendered 1 page/i);
+			expect(code).toBe(0);
+
+			const index = await fs.readFile(path.join(env.tmp.path, 'dist', 'index.html'), 'utf8');
+			expect(index).toMatch('it works');
+		});
+
+		it('should inject prerendered data into the html', async () => {
+			await loadFixture('prerender-data', env);
+			instance = await runWmr(env.tmp.path, 'build', '--prerender');
+			const code = await instance.done;
+			expect(instance.output.join('\n')).toMatch(/Prerendered 1 page/i);
+			expect(code).toBe(0);
+
+			const index = await fs.readFile(path.join(env.tmp.path, 'dist', 'index.html'), 'utf8');
+			expect(index).toMatch('<script type="isodata">{"hello":"world"}</script>');
+		});
+
 		it('should support prerendered HTML, title & meta tags', async () => {
 			await loadFixture('prod-head', env);
 			instance = await runWmr(env.tmp.path, 'build', '--prerender');
@@ -357,15 +550,15 @@ describe('production', () => {
 
 			const index = await fs.readFile(path.join(env.tmp.path, 'dist', 'index.html'), 'utf8');
 			expect(index).toMatch('<title>Page: /</title>');
-			expect(index).toMatch('<link rel="icon" href="data:,favicon-for-/">');
+			expect(index).toMatch('<link href="data:,favicon-for-/" rel="icon">');
 			expect(index).toMatch('<h1>page = /</h1>');
-			expect(index).toMatch(`<meta property="og:title" content="Become an SEO Expert">`);
+			expect(index).toMatch(`<meta content="Become an SEO Expert" property="og:title">`);
 
 			const other = await fs.readFile(path.join(env.tmp.path, 'dist', 'other.html'), 'utf8');
 			expect(other).toMatch('<title>Page: /other.html</title>');
-			expect(other).toMatch('<link rel="icon" href="data:,favicon-for-/other.html">');
+			expect(other).toMatch('<link href="data:,favicon-for-/other.html" rel="icon">');
 			expect(other).toMatch('<h1>page = /other.html</h1>');
-			expect(other).toMatch(`<meta property="og:title" content="Become an SEO Expert">`);
+			expect(other).toMatch(`<meta content="Become an SEO Expert" property="og:title">`);
 		});
 
 		it('should support prerendering json', async () => {